# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Trade Analytics Copilot API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /trades:
    get:
      operationId: listTrades
      tags: [trades]
      summary: List all trades
      parameters:
        - name: symbol
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [open, closed, all]
        - name: setup
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Trade list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TradeList"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /trades/{id}:
    get:
      operationId: getTrade
      tags: [trades]
      summary: Get a single trade
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Trade detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TradeDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  responses:
    Unauthorized:
      description: Not authenticated
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
...
    Trade:
      type: object
      properties:
        id:
          type: string
        symbol:
          type: string
        side:
          type: string
          enum: [long, short]
        status:
          type: string
          enum: [open, closed]
        setup:
          type: string
        entryPrice:
          type: number
        exitPrice:
          type: number
        quantity:
          type: number
        pnl:
          type: number
        pnlPercent:
          type: number
        rr:
          type: number
        openTime:
          type: string
        closeTime:
          type: string
        broker:
          type: string
        tags:
          type: array
          items:
            type: string
      required: [id, symbol, side, status, setup, entryPrice, quantity, pnl, openTime, broker, tags]
    TradeList:
      type: object
      properties:
        trades:
          type: array
          items:
            $ref: "#/components/schemas/Trade"
        total:
          type: integer
      required: [trades, total]
    TradeDetail:
      allOf:
        - $ref: "#/components/schemas/Trade"
        - type: object
          properties:
            notes:
              type: string
            screenshots:
              type: array
              items:
                type: string
            stopLoss:
              type: number
            takeProfit:
              type: number
            mae:
              type: number
            mfe:
              type: number
            commission:
              type: number
...